refactor: gateway runtime structure
