+++ jina/parsers/hubble/push.py
@@ -1,6 +1,7 @@
 """Argparser module for hub push"""
 
 import argparse
+
 from ..helper import add_arg_group
