+++ jina/parsers/optimizer.py
@@ -1,7 +1,8 @@
 """Argparser module for Optimizer"""
+from typing import Optional
+
 from .base import set_base_parser
 from .peapods.base import mixin_base_ppr_parser
-from typing import Optional
 
 # noinspection PyUnreachableCode
 if False:
