+++ jina/peapods/runtimes/__init__.py
@@ -5,9 +5,9 @@ def list_all_runtimes():
     # noqa: DAR201
     """
     from ...peapods.runtimes.base import BaseRuntime
-    from .asyncio.grpc import GRPCRuntime
-    from .asyncio.http import HTTPRuntime
-    from .asyncio.websocket import WebSocketRuntime
+    from .gateway.grpc import GRPCRuntime
+    from .gateway.http import HTTPRuntime
+    from .gateway.websocket import WebSocketRuntime
     from .container import ContainerRuntime
     from .jinad import JinadRuntime
     from .zmq.zed import ZEDRuntime
@@ -26,9 +26,9 @@ def get_runtime(name: str):
     # noqa: DAR201
     """
     from ...peapods.runtimes.base import BaseRuntime
-    from .asyncio.grpc import GRPCRuntime
-    from .asyncio.http import HTTPRuntime
-    from .asyncio.websocket import WebSocketRuntime
+    from .gateway.grpc import GRPCRuntime
+    from .gateway.http import HTTPRuntime
+    from .gateway.websocket import WebSocketRuntime
     from .container import ContainerRuntime
     from .jinad import JinadRuntime
     from .zmq.zed import ZEDRuntime
