+++ jina/peapods/runtimes/gateway/websocket/app.py
@@ -55,7 +55,7 @@ def get_fastapi_app(args: 'argparse.Namespace', logger: 'JinaLogger'):
                 yield Request(data)
 
         try:
-            async for msg in servicer.Call(request_iterator=req_iter(), context=None):
+            async for msg in servicer.Call(request_iterator=req_iter()):
                 await websocket.send_bytes(msg.binary_str())
         except WebSocketDisconnect:
             manager.disconnect(websocket)
