+++ jina/peapods/runtimes/jinad/__init__.py
@@ -1,13 +1,13 @@
-import time
-import asyncio
 import argparse
+import asyncio
+import time
 from typing import Optional
 
-from ....enums import RemoteWorkspaceState
-from ..asyncio.base import AsyncZMQRuntime
-from ....helper import cached_property, colored
-from ....excepts import DaemonConnectivityError
 from .client import PeaDaemonClient, WorkspaceDaemonClient
+from ..zmq.asyncio import AsyncZMQRuntime
+from ....enums import RemoteWorkspaceState
+from ....excepts import DaemonConnectivityError
+from ....helper import cached_property, colored
 
 
 class JinadRuntime(AsyncZMQRuntime):
