+++ jina/peapods/runtimes/zmq/zed.py
@@ -32,9 +32,13 @@ from ....types.routing.table import RoutingTable
 class ZEDRuntime(ZMQRuntime):
     """Runtime procedure leveraging :class:`ZmqStreamlet` for Executor."""
 
-    def __init__(self, args: 'argparse.Namespace', ctrl_addr: str):
+    def __init__(self, args: argparse.Namespace, ctrl_addr: str):
+        """Initialize private parameters and execute private loading functions.
+
+        :param args: args from CLI
+        :param ctrl_addr: control port address
+        """
         super().__init__(args, ctrl_addr)
-        """Initialize private parameters and execute private loading functions."""
         self._id = random_identity()
         self._last_active_time = time.perf_counter()
